refactor: articleForm 삭제 후 WriteArticlePage에 병합
